UBUNTU: SAUCE: Add quirk for ASUS Z37E to make sound audible after resume
Bug: #25896
UpstreamStatus: not yet merged

Adds a quirk for SSID 1043:8284 to use 6stack-dig model by default so
that sound is audible after resuming.

Test Case: Off a fresh boot, resuming results in no sound until the
sound modules are unloaded and then snd_hda_intel is reloaded.  Because
there is no quirk entry for this model, the BIOS autoconfig is used,
and the autoconfig fails to write back the correct information.

Behavior with fix applied: Using the 6stack-dig quirk, sound is audible
after resuming.

Signed-off-by: Mackenzie Morgan <maco.m@ubuntu.com>
Signed-off-by: Daniel T Chen <crimsun@ubuntu.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>