UBUNTU: SAUCE: (revert before 2.6.28.y update) [PATCH] ext4: tone down ext4_da_writepages warnings
If the filesystem has errors, ext4_da_writepages() will return a *lot*
of errors, including lots and lots of stack dumps.  While it's true
that we are dropping user data on the floor, which is unfortunate, the
stack dumps aren't helpful, and they tend to obscure the true original
root cause of the problem.  So in the case where the filesystem has
aborted, return an EROFS right away.

Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>
(cherry picked from commit 2a21e37e48b94388f2cc8c0392f104f5443d4bb8)
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>