UBUNTU: SAUCE: iwlagn -- fix rfkill when on when driver loaded
OriginalAuthor: Mohamed Abbas <mohamed.abbas@intel.com>
Bug: #193970

When the iwlagn driver is loaded and the rfkill switch is already on
then we do not start notifications.  This prevents the switch enabling
wireless.

This patch is currently on its way upstream.  This has been tested on an
iwl5300.  Proposing this fix for SRU for Intrepid. Carry as a SAUCE patch
until superseded.

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>