ricoh_mmc: Handle newer models of Ricoh controllers
The latest generation of laptops are shipping with a newer
model of Ricoh chip where the firewire controller is the
primary PCI function but a cardbus controller is also present.

The existing code assumes that if a cardbus controller is,
present, then it must be the one to manipulate - but the real
rule is that you manipulate PCI function 0. This patch adds an
additional constraint that the target must be function 0.

Signed-off-by: Philip Langdale <philipl@overt.org>
Signed-off-by: Pierre Ossman <drzeus@drzeus.cx>
(cherry picked from commit 0527a60c2b6bd7ab20e82cc5e488659e20eaaacd)

Signed-off-by: Tim Gardner <tim.gardner@canonical.com>