UBUNTU: update the templates so that we have spaces following the title line
git tends to missrepresents the tag fields such as 'Bug: #' appending
them to the title line if there is no blank line after it.  Add these
blank lines to the templates to encourge them to be in the final
commits.

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Tim Gardener <tim.gardner@canonical.com>
(cherry picked from commit a0b081fe608727148828776159abc09852b266c4)