UBUNTU: synchronise our linux-libc-dev with the kernel userspace headers
Bug: #300803

It seems that the kernel user space headers have expanded to include
some new directories for sound and the like, the full list:

    asm asm-generic linux mtd rdma sound video

We currently explicitly list the headers we will include in our
linux-libc-dev.  Which means we only include the asm* and linux
directories, which makes difficulties for porters.  Debian already
includes all of these in their packages.

It seems appropriate that we include all of the officially exported
userspace headers from the kernel.  Move to taking all of the headers that
the header_install target generates so that we track and furture changes.

Also take this oppotunity to elide a couple of build file which seem to
be being produced, .install and ..install.cmd which seem to be noise.

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
(cherry picked from commit 4805f9cea513e6e2cdda6a56720114d0e124605d)