Staging: android: binder: fix build errors
This fixes the build errors and warnings in the binder driver.  It can't
be a module, due to a lack of some of the symbols being exported.

Also added a MODULE_LICENSE(), as it was missing.

Cc: Arve Hjønnevåg <arve@android.com>
Cc: Brian Swetland <swetland@google.com>
Cc: Robert Love <rlove@google.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>
(cherry picked from commit cbb3b89b9c27b4edafc20fb022ee1ec61d1471e4)