UBUNTU: Enable dm-raid4-5
Bug: #309378

The dm-raid4-5 driver was disabled as it did not cleanly compile when
rebased to 2.6.28.  This was due to two errors:

1) the removal of {set,clear}_page_locked() as the atomic versions were
   no longer used elsewhere in the kernel, and
2) the array_too_big() function was renamed to dm_array_too_big() to
   allow it to be exported.

As raid is a highly critical subsystem I have taken the safest possible
approach and assumed that RAID does require the atomic flag modifiers.
This patch therefore reinstates {set,clear}_page_locked() and moves the
code to using the new dm_array_too_big() interface.  Finally it reenabled
the raid support.

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>